Added range queries, ext. node iterators, refact.
Extended the node iterators to be able to start at any given
hypercube address. Depending on the node type the first valid
address is searched.

Refactoring:
- moved nodes, utils and iterators into seperate subfolders
- moved bit handling operations from the node class into a utility
class and added several more helper methods to reuse common
functionality (e.g. converting between bitsets and values)
- values are now handled in (unsigned) long instead of int
- the node iterators now work like standard C++ iterators
for (it = node.begin(); it != node.end(); ++it) instead of
the previous < comparison
--> now they jump from valid address to valid address and stop
at the biggest possible address + 1 which is per definition (2^dim)
--> also the LHC node does not need to manage the highest
contained address any more

Fixes:
- the AHC node iterator did not the the address

Range queries:
- the lookup method of a node can now return a stack of all
visited nodes during the lookup (this is used in range queries)
- added range queries to the PH-Tree and the node class by first
performing a lookup, initializing an iterator with the visited
nodes and then returning the iterator
- added a range query iterator which takes a node sequence (from a
lookup operation) and then iterates over contained entries within
the specified range given by the lower left and upper right corner
The iteration works as follows:
1. descend to the node containing (or failing to contain) the lower
left node --> point query
2. iterate using definied bit masks over valid entries inside the node
while descending into contained subnodes and ascending in the tree if
the end of the current node is reached (i.e. the upper valid bit mask)
3. if the end of the root node is reached, there are no more entries in
the range
--> for ascending and descending need to manage a stack that holds the
current valid address and the currently processed node

+++ src/Entry.h
@@ -18,7 +18,7 @@ class Entry {
 
 public:
 
-	Entry(std::vector<int> values, int bitLength);
+	Entry(std::vector<long> values, int bitLength);
 	Entry(std::vector<std::vector<bool>> values);
 	~Entry();
